+++ src/context.ts
@@ -26,6 +26,7 @@ import { SystemPromptBuilder } from './system-prompt-builder';
 import { aisdk } from './utils/ai-sdk';
 import { getEnv } from './utils/env';
 import { getGitStatus } from './utils/git';
+import { ModelInfo } from './utils/model';
 import { relativeToHome } from './utils/path';
 
 type Env = {
@@ -55,9 +56,13 @@ type Paths = {
   projectConfigDir: string;
 };
 
+type ArgvConfig = Partial<Config> & {
+  appendSystemPrompt?: string;
+};
+
 export interface CreateContextOpts {
   cwd: string;
-  argvConfig?: Partial<Config>;
+  argvConfig?: ArgvConfig;
   productName?: string;
   version?: string;
   plugins?: Plugin[];
@@ -69,7 +74,7 @@ export class Context {
   productName: string;
   version: string;
   config: Config;
-  argvConfig: Partial<Config>;
+  argvConfig: ArgvConfig;
   pluginManager: PluginManager;
   mcpManager: MCPManager;
   mcpTools: Tool[];
@@ -80,6 +85,7 @@ export class Context {
   paths: Paths;
   slashCommands: SlashCommandRegistry;
   env: Env;
+  modelInfo: ModelInfo;
   constructor(opts: ContextOpts) {
     this.cwd = opts.cwd;
     this.productName = opts.productName || PRODUCT_NAME;
@@ -96,6 +102,7 @@ export class Context {
     this.paths = opts.paths;
     this.slashCommands = opts.slashCommands;
     this.env = opts.env;
+    this.modelInfo = new ModelInfo(this);
   }
 
   static async create(opts: CreateContextOpts) {
@@ -171,6 +178,7 @@ async function createContext(opts: CreateContextOpts): Promise<Context> {
   const apply = async (hookOpts: any) => {
     return pluginManager.apply({ ...hookOpts, pluginContext: tempContext });
   };
+
   const tempContext = {
     ...opts,
     pluginManager,
@@ -184,6 +192,7 @@ async function createContext(opts: CreateContextOpts): Promise<Context> {
     memo: initialConfig,
     type: PluginHookType.SeriesMerge,
   });
+
   debug('resolvedConfig', resolvedConfig);
   tempContext.config = resolvedConfig;
   await apply({
@@ -191,6 +200,16 @@ async function createContext(opts: CreateContextOpts): Promise<Context> {
     args: [{ resolvedConfig }],
     type: PluginHookType.Series,
   });
+
+  // Placed after config so we can customize prompts based on the user's model
+  opts.argvConfig = await apply({
+    hook: 'argvConfig',
+    args: [{}],
+    memo: opts.argvConfig,
+    type: PluginHookType.SeriesMerge,
+  });
+
+  debug('argvConfig', opts.argvConfig);
 
   const env = await getEnv();
   const generalInfo = await apply({