+++ src/plugin.ts
@@ -175,4 +175,6 @@ export type Plugin = {
     },
   ) => Promise<any> | any;
   command?: (this: PluginContext) => Promise<any[]> | any[];
+  argvConfig?: (this: PluginContext) => Promise<any> | any;
+  modelInfo?: (this: PluginContext) => Promise<any> | any;
 };